requirements.txt
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
flask>=2.3.0
Pillow>=9.5.0
requests>=2.28.0